backend/requirements.txt
fastapi[standard]
uvicorn
pydantic
google-genai
python-dotenv
